Move throttler tests (#7318)
